test(AuthControllerTest): 인증 컨트롤러 테스트 코드 추가
