+++ src/test/java/com/sprint/mission/discodeit/controller/ChannelControllerTest.java
@@ -23,7 +23,7 @@ import java.util.UUID;
 
 import static com.sprint.mission.discodeit.util.mock.channel.ChannelInfo.CHANNEL_NAME;
 import static com.sprint.mission.discodeit.util.mock.channel.ChannelInfo.UPDATED_CHANNEL_NAME;
-import static com.sprint.mission.discodeit.util.mock.user.SetUpUserInfo.OTHER_USER;
+import static com.sprint.mission.discodeit.util.mock.user.UserInfo.OTHER_USER;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.when;
