+++ src/test/java/com/sprint/mission/discodeit/util/mock/user/UserInfo.java
@@ -1,0 +1,28 @@
+package com.sprint.mission.discodeit.util.mock.user;
+
+public enum UserInfo {
+    LOGIN_USER("황지환", "hwang@naver.com", "12345"),
+    OTHER_USER("박지환", "park@naver.com", "12345");
+
+    private final String name;
+    private final String email;
+    private final String password;
+
+    UserInfo(String name, String email, String password) {
+        this.name = name;
+        this.email = email;
+        this.password = password;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public String getEmail() {
+        return email;
+    }
+
+    public String getPassword() {
+        return password;
+    }
+}
